backend/requirements.txt
fastapi[standard]
langgraph
supabase
groq
python-dotenv
